Crated a viewModel for the SingleMaterialSummary and the wrappers to ensure it is displayed correctly

+++ app/src/main/java/com/example/myapplication/ui/screen/Select/SelectActivity.kt
@@ -85,7 +85,7 @@ fun SelectActivity(
                         actions.setChecked(item.id)
                     },
                     onClick = {
-                        navController.navigate(NavigationRoute.SingleMaterialSummary(item.id))
+                        navController.navigate(NavigationRoute.SingleMaterialSummary(item.id.toInt()))
                     },
                     type = item.type.toString()
                 )
